app/requirements.txt
alembic==0.9.6
bcrypt==3.1.2
boto==2.48.0
boto3==1.5.31
botocore==1.8.45
certifi==2018.1.18
aniso8601==1.3.0
click==6.7
coverage==4.4.2
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Cache==0.13.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
nose==1.3.7
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2
requests
gunicorn==19.7.1
cherrypy
PyJWT==1.4.2
sendgrid
coveralls
psycopg2
s3transfer==0.1.13
tinys3==0.1.12
xlrd==1.1.0
xlwt==1.3.0